+++ src/shared/ui/gatheringCard/GatheringCard.tsx
